Create AbstractPrideSlider
